pyproject.toml
[project]
name = "parade-state-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.169.0",
    "loguru>=0.7.3",
    "pandas>=2.2.3",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.1",
]
